ampl-growth-marketing-agent/agent/requirements.txt
google-ads==24.1.0
python-dotenv==1.0.1
boto3==1.34.0
requests==2.31.0
